requirements.txt
flask==2.3.3
flask-cors==4.0.0
groq==0.4.1
faiss-cpu==1.7.4
numpy==1.26.0
scikit-learn==1.3.1
rank-bm25==0.2.2
langdetect==1.0.9
sentence-transformers==2.2.2
nltk==3.8.1
python-dotenv==1.0.0